API/requirements.txt
black==25.1.0
Django==4.2.20
djangorestframework==3.16.0
python-docx==1.1.2
docx2pdf==0.1.8
django-cors-headers==3.14.0